лады

+++ StrangeServerCSharp/Clan.cs
@@ -14,18 +14,21 @@ namespace StrangeServerCSharp
         {
             memberList = "";
         }
+
         public static void InitClans()
         {
             if (BDClass.THIS.clans.Count() == 0)
             {
-                for (int i = 1; i <= 219; i++)
+                for (var i = 1; i <= 219; i++)
                 {
-                    BDClass.THIS.clans.Add(new Clan() { abr = "",owner = "", name = ""});
+                    BDClass.THIS.clans.Add(new Clan() { abr = "", owner = "", name = "" });
                 }
             }
+
             BDClass.THIS.SaveChanges();
         }
-        public static void CreateClan(int id,string name, string abr, Player p)
+
+        public static void CreateClan(int id, string name, string abr, Player p)
         {
             var clan = BDClass.THIS.clans.First(i => i.id == id);
             clan.name = name;
@@ -36,6 +39,7 @@ namespace StrangeServerCSharp
             p.SendClan();
             BDClass.THIS.SaveChanges();
         }
+
         public void AddReq(int id)
         {
             if (!reqs.Contains(id))
@@ -43,6 +47,7 @@ namespace StrangeServerCSharp
                 reqs.Add(id);
             }
         }
+
         public void RemoveReq(int id)
         {
             if (reqs.Contains(id))
@@ -50,27 +55,31 @@ namespace StrangeServerCSharp
                 reqs.Remove(id);
             }
         }
+
         public static string winid = "!!clan";
         public List<int> reqs = new List<int>();
-        public static void Open(string t,Player p)
+
+        public static void Open(string t, Player p)
         {
-            var c = new HorbConst();
+            var builder = new HorbBuilder();
             if (t == winid)
             {
                 if (p.clanid == 0)
                 {
                     return;
                 }
+
                 var clan = BDClass.THIS.clans.First(clan => clan.id == p.clanid);
-                c.AddTitle(clan.name);
-                c.AddCard("c", p.clanid.ToString(), $"<color=white>Winx </color>\nУчастники: <color=white> {1 + clan.GetMemberList().Count} </color> ");
-                c.AddTab("ОБЗОР", ""); c.AddTab("СПИСОК", "list");
-                c.AddButton("ПОКИНУТЬ КЛАН", "leave");
+                builder.SetTitle(clan.name)
+                    .AddCard("c", p.clanid.ToString(),
+                        $"<color=white>Winx </color>\nУчастники: <color=white> {1 + clan.GetMemberList().Count} </color> ")
+                    .AddTab("ОБЗОР", "");
+                builder.AddTab("СПИСОК", "list");
+                builder.AddButton("ПОКИНУТЬ КЛАН", "leave");
                 if (clan.reqs.Count > 0)
                 {
-                   c.AddTab($"<color=#ff8888ff>ЗАЯВКИ ({clan.reqs.Count})</color>", "reqs");
+                    builder.AddTab($"<color=#ff8888ff>ЗАЯВКИ ({clan.reqs.Count})</color>", "reqs");
                 }
-
             }
             else if (t.StartsWith("reqs"))
             {
@@ -78,20 +87,23 @@ namespace StrangeServerCSharp
                 {
                     return;
                 }
+
                 var clan = BDClass.THIS.clans.First(clan => clan.id == p.clanid);
-                c.AddTitle(clan.name);
-                c.AddList();
-                int num = 0;
-                c.AddTab("ОБЗОР", "!!clan"); c.AddTab("СПИСОК", "list");
+                builder.SetTitle(clan.name);
+                var num = 0;
+                builder.AddTab("ОБЗОР", "clan");
+                builder.AddTab("СПИСОК", "list");
+
                 if (clan.reqs.Count > 0)
                 {
-                    c.AddTab($"<color=#ff8888ff>ЗАЯВКИ ({clan.reqs.Count})</color>", "");
+                    builder.AddTab($"<color=#ff8888ff>ЗАЯВКИ ({clan.reqs.Count})</color>", "");
                 }
+
                 foreach (var id in clan.reqs)
                 {
                     var player = BDClass.THIS.players.First(i => i.id == id);
                     num++;
-                    c.AddListLine($"{num} <color=white>{player.name}</color>", "...", $"req:{id}:{clan.id}");
+                    builder.AddListLine($"{num} <color=white>{player.name}</color>", "...", $"req:{id}:{clan.id}");
                 }
             }
             else if (t.StartsWith("listrow"))
@@ -105,10 +117,9 @@ namespace StrangeServerCSharp
                 {
                     return;
                 }
-                c.AddTitle(clan.name);
-                c.AddList();
+                builder.SetTitle(clan.name);
                 int num = 0;
-                c.AddTab("ОБЗОР", "!!clan"); c.AddTab("СПИСОК", "list");
+                builder.AddTab("ОБЗОР", "clan"); builder.AddTab("СПИСОК", "list");
                 if (!int.TryParse(t.Split(':')[1],out var v))
                 {
                     return;
@@ -120,10 +131,10 @@ namespace StrangeServerCSharp
                 }
                 catch (Exception ex) { }
                 if (player == null) { return; }
-                c.SetText($"@@ПРОФИЛЬ СОКЛАНА\n\nНик:<color=white>{player.name}</color>\nУровень:{player.lvl.Sum()}");
+                builder.SetText($"@@ПРОФИЛЬ СОКЛАНА\n\nНик:<color=white>{player.name}</color>\nУровень:{player.lvl.Sum()}");
                 if (p.id != player.id && player.id.ToString() != clan.owner)
                 {
-                    c.AddButton("ИСКЛЮЧИТЬ ИЗ КЛАНА", "listrow_kick:" + player.id);
+                    builder.AddButton("ИСКЛЮЧИТЬ ИЗ КЛАНА", "listrow_kick:" + player.id);
                 }
 
             }
@@ -133,27 +144,31 @@ namespace StrangeServerCSharp
                 {
                     return;
                 }
+
                 var clan = BDClass.THIS.clans.First(clan => clan.id == p.clanid);
-                c.AddTitle(clan.name);
-                c.AddClanList();
-                int num = 1;
-                c.AddTab("ОБЗОР", "!!clan"); c.AddTab("СПИСОК", "");
+                builder.SetTitle(clan.name);
+                builder.AddClanList();
+                var num = 0;
+                builder.AddTab("ОБЗОР", "clan");
+                builder.AddTab("СПИСОК", "");
                 if (clan.reqs.Count > 0)
                 {
-                    c.AddTab($"<color=#ff8888ff>ЗАЯВКИ ({clan.reqs.Count})</color>", "");
+                    builder.AddTab($"<color=#ff8888ff>ЗАЯВКИ ({clan.reqs.Count})</color>", "");
                 }
                 var player = BDClass.THIS.players.First(i => i.id.ToString() == clan.owner);
-                c.AddClanListLine("0", $"{num}.<color=white>{player.name}</color>", "", "listrow:" + clan.owner);
+                builder.AddClanListLine("0", $"{num}.<color=white>{player.name}</color>", "", "listrow:" + clan.owner);
                 foreach (var id in clan.GetMemberList())
                 {
                     player = BDClass.THIS.players.First(i => i.id == id);
                     num++;
-                    c.AddClanListLine("0", $"{num}.<color=white>{player.name}</color>", "", "listrow:" + id);
+                    builder.AddClanListLine("0", $"{num}.<color=white>{player.name}</color>", "", "listrow:" + id);
                 }
             }
-            c.Send("!!clan." + t, p);
+
+            builder.Send("!!clan." + t, p);
         }
-        public static void DeleteClan(int id,Player p)
+
+        public static void DeleteClan(int id, Player p)
         {
             var clan = BDClass.THIS.clans.First(i => i.id == id);
             clan.name = "";
@@ -165,37 +180,50 @@ namespace StrangeServerCSharp
             p.SendMoney();
             BDClass.THIS.SaveChanges();
         }
+
         public static Clan finclan(int id)
         {
             return BDClass.THIS.clans.First(i => i.id == id);
         }
+
         public void Addmember(int memberid)
         {
             Player player = null;
             try
             {
                 player = BDClass.THIS.players.First(p => p.id == memberid);
-                    } catch (Exception ex) { }
-            if (player == null) { return;  }
+            }
+            catch (Exception ex)
+            {
+            }
+
+            if (player == null)
+            {
+                return;
+            }
+
             player.clanid = id;
             player.SendClan();
             if (!GetMemberList().Contains(memberid))
             {
                 memberList += ":" + memberid.ToString();
             }
+
             BDClass.THIS.SaveChanges();
         }
+
         public static List<Clan> GetAvlClanIcon()
         {
-            List<Clan> l = new List<Clan>();
+            var l = new List<Clan>();
             var col = 8;
-            for (int i = 0; i < col; i++)
+            for (var i = 0; i < col; i++)
             {
-                int r = World.Random.Next(1, 219);
+                var r = World.Random.Next(1, 219);
                 var c = BDClass.THIS.clans.Where(j => string.IsNullOrEmpty(j.owner)).ToList();
                 var clan = c.ElementAt(new Index(World.Random.Next(1, c.Count())));
                 l.Add(clan);
             }
+
             return l;
         }
         public void Kick(int id)
@@ -207,9 +235,13 @@ namespace StrangeServerCSharp
             }
             else
             {
-                if (owner == id.ToString())
+                if (owner == id.ToString() && members.Count < 1)
                 {
-                    DeleteClan(this.id,BDClass.THIS.players.First(p => p.id == id));
+                    DeleteClan(this.id, BDClass.THIS.players.First(p => p.id == id));
+                }
+                else
+                {
+                    return;
                 }
             }
             Player player = null;
@@ -231,14 +263,19 @@ namespace StrangeServerCSharp
         }
         public List<int> GetMemberList()
         {
-            string[] m = memberList.Split(":");
-            int[] mem = new int[] { };
-            foreach (string mid in m)
+            var m = memberList.Split(":");
+            var mem = new int[] { };
+            foreach (var mid in m)
             {
-                if (!string.IsNullOrEmpty(mid)) { mem = mem.Concat(new int[] { int.Parse(mid) }).ToArray(); }
+                if (!string.IsNullOrEmpty(mid))
+                {
+                    mem = mem.Concat(new int[] { int.Parse(mid) }).ToArray();
+                }
             }
+
             return mem.ToList();
         }
+
         public string abr { get; set; }
         public int id { get; set; }
         public string name { get; set; }
